swap kube for csswizardry
I know I'm all over the place on this one, but those grids just didn't
feel right to me.  This is a touch more css, but I think it will be
better?  I don't know.  Probably best to just go with foundation, but
I'm going to give this a shot and see how it goes.

[#IhavenoideawhatImdoing]